api/pyproject.toml
[tool.poetry]
name = "flask-diagram-generator"
version = "0.1.0"
description = "A Flask service that generates architecture diagrams from CloudFormation templates"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
diagrams = "^0.23.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
